backend/requirements.txt
Django>=4.2,<5.1
djangorestframework>=3.14
django-cors-headers>=4.3
gunicorn>=21.2

# GIS dependencies (shared with engines)
geopandas>=0.14
pandas>=2.0
numpy>=1.24
shapely>=2.0
rasterio>=1.3
rasterstats>=0.19
fiona>=1.9
networkx>=3.1
pyproj>=3.6
psutil>=5.9
